Do not nest the build output
When compiling with Visual Studio < 2019, CMake used to nest the build
output under `x64/vc15/`. This does not happen anymore in VS 2019,
so the build output was not being copied to the final directory.

This fix disables nesting for all versions and updates the paths to the
built files accordingly.